Move group membership views into a separate file
Group membership views and group views are not really the same thing and
it's not very nice to group them all together in one big file. We
already have a separate `services/group_members.py` and a separate
`traversal/group_membership.py`, move the group membership views into
a separate `views/api/group_members.py` as well.
