+++ tests/unit/h/views/api/groups_test.py
@@ -1,13 +1,12 @@
 from unittest.mock import PropertyMock, call, create_autospec, sentinel
 
 import pytest
-from pyramid.httpexceptions import HTTPConflict, HTTPNoContent, HTTPNotFound
+from pyramid.httpexceptions import HTTPConflict
 
 import h.views.api.groups as views
-from h import presenters
-from h.models import GroupMembership, User
+from h.models import User
 from h.schemas.base import ValidationError
-from h.traversal import GroupContext, GroupMembershipContext
+from h.traversal import GroupContext
 from h.views.api.exceptions import PayloadError
 
 
@@ -403,73 +402,6 @@ class TestUpdate:
         return pyramid_request
 
 
-class TestReadMembers:
-    def test_it(self, context, pyramid_request, UserJSONPresenter):
-        context.group.members = [sentinel.member_1, sentinel.member_2]
-        presenter_instances = UserJSONPresenter.side_effect = [
-            create_autospec(presenters.UserJSONPresenter, instance=True, spec_set=True),
-            create_autospec(presenters.UserJSONPresenter, instance=True, spec_set=True),
-        ]
-
-        response = views.read_members(context, pyramid_request)
-
-        assert UserJSONPresenter.call_args_list == [
-            call(sentinel.member_1),
-            call(sentinel.member_2),
-        ]
-        presenter_instances[0].asdict.assert_called_once_with()
-        presenter_instances[1].asdict.assert_called_once_with()
-        assert response == [
-            presenter_instances[0].asdict.return_value,
-            presenter_instances[1].asdict.return_value,
-        ]
-
-
-class TestRemoveMember:
-    def test_it(self, context, pyramid_request, group_members_service):
-        response = views.remove_member(context, pyramid_request)
-
-        group_members_service.member_leave.assert_called_once_with(
-            context.group, context.user.userid
-        )
-        assert isinstance(response, HTTPNoContent)
-
-    @pytest.fixture
-    def context(self, factories):
-        group = factories.Group.build()
-        user = factories.User.build()
-        membership = GroupMembership(group=group, user=user)
-        return GroupMembershipContext(group=group, user=user, membership=membership)
-
-
-@pytest.mark.usefixtures("group_members_service")
-class TestAddMember:
-    def test_it(self, pyramid_request, group_members_service, context):
-        response = views.add_member(context, pyramid_request)
-
-        group_members_service.member_join.assert_called_once_with(
-            context.group, context.user.userid
-        )
-        assert isinstance(response, HTTPNoContent)
-
-    def test_it_with_authority_mismatch(self, pyramid_request, context):
-        context.group.authority = "other"
-
-        with pytest.raises(HTTPNotFound):
-            views.add_member(context, pyramid_request)
-
-    @pytest.fixture
-    def pyramid_request(self, pyramid_request):
-        pyramid_request.matchdict = {"userid": sentinel.userid}
-        return pyramid_request
-
-    @pytest.fixture
-    def context(self, factories):
-        group = factories.Group.build()
-        user = factories.User.build(authority=group.authority)
-        return GroupMembershipContext(group=group, user=user, membership=None)
-
-
 @pytest.fixture
 def assert_it_returns_group_as_json(GroupJSONPresenter, pyramid_request):
     def assert_it_returns_group_as_json(
@@ -509,13 +441,6 @@ def GroupJSONPresenter(mocker):
 def GroupsJSONPresenter(mocker):
     return mocker.patch(
         "h.views.api.groups.GroupsJSONPresenter", autospec=True, spec_set=True
-    )
-
-
-@pytest.fixture(autouse=True)
-def UserJSONPresenter(mocker):
-    return mocker.patch(
-        "h.views.api.groups.UserJSONPresenter", autospec=True, spec_set=True
     )
